Add configuration local app data support
The SearchConfiguration.xml file and search indexes were saved directly
at the location of the executing assembly, which meant that different
configuration and indexes were used when we ran the GUI Viewer vs the
CLI application or when we ran Debug vs Release. It also meant data
files were saved along with the binaries.

The new behavior first checks if there is a configuration file at the
location of the executing assembly and uses it if that's the case. That
allows backward compatibility for users that have their configuration at
that location. However, if there is no configuration at that location,
we now look into `%LocalAppData%/FlashSearcher` and save the default
configuration there if there's isn't one already. That allows sharing a
single configuration between GUI/CLI and Debug/Release, and saves user
data in her/his own folder.